pyproject.toml
[project]
name = "rpg_agent"
version = "0.0.1"
description = "An agent that helps you play Divinity Original Sin II."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "instructor>=1.6.3",
    "mistralai>=1.2.1",
    "mss>=9.0.2",
    "openai>=1.54.3",
    "opencv-python>=4.10.0.84",
    "pip>=24.2",
    "pillow>=10.4.0",
    "PyAutoGUI>=0.9.54",
    "pynput>=1.7.7",
    "pyobjc-framework-Quartz>=10.3.1",
    "python-dotenv>=1.0.1",
    "PyWinCtl>=0.4.1",
    "weave>=0.51.19",
    "uv>=0.4.17",
    "black>=24.8.0",
    "isort>=5.13.2",
    "pytest>=8.3.3",
    "ruff>=0.6.8",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]